packages/py/tracing/pyproject.toml
[project]
name = "uselemma-tracing"
version = "1.0.0"
description = "OpenTelemetry-based tracing module for Lemma"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
keywords = ["opentelemetry", "tracing", "observability", "monitoring", "llm", "instrumentation"]
dependencies = [
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp-proto-http>=1.28.0",
]

[project.optional-dependencies]
openai-agents = ["openinference-instrumentation-openai-agents>=0.1.0"]

[project.urls]
Homepage = "https://github.com/uselemma/tracing"
Repository = "https://github.com/uselemma/tracing"
Issues = "https://github.com/uselemma/tracing/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/uselemma_tracing"]

[tool.hatch.build.targets.sdist]
include = ["src/", "LICENSE", "README.md"]
